Move :class: ComparisonBundle to rmgpy/tools/plot.py
Moved to make more consistent and avoid too many reflective imports: plot.py imports data.py, but not the other way around. Eventually I will want ComparisonBundle to have a .plot() function
